Implement new, vastly more efficient renderer
The first change is that it is now possible to redraw only a single chunk, instead of a whole .mca region. This is implemented by loading the existing image file and only redrawing the chunk. Previously, the images where stored as JPG files, so loading them, changing something and then saving them again was not an option.

Currently, the renderer runs at full speed; slowing it down still needs to be reimplemented.

The second change is that zooming out is now possible. This is implemented by a tree-like data structure: every "super region" holds 2 by 2 smaller regions. These regions again hold smaller regions, etcetera, until we end up at the scale of the normal .mca regions. The renderer starts updating at the chunk-level, and once all updates in a region are finished, the super region is redrawn.